pyproject.toml
[tool.poetry]
name = "simpledjangoproject"
version = "0.1.0"
description = ""
authors = ["Igor Lagutin <igor.lagutin@pythagoras.technology>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
gunicorn = "^22.0.0"
boto3 = "^1.34.149"
django-storages = "^1.14.4"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
